pyproject.toml
[tool.poetry]
name = "pse"
version = "0.1.0"
description = ""
authors = ["Danie Marques de Almeida <danielmarques.enfermagem@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
geopandas = "^1.0.1"
folium = "^0.17.0"
streamlit = "^1.39.0"
plotly = "^5.24.1"
ruff = "^0.6.9"
black = "^24.8.0"
taskipy = "^1.13.0"
python-dotenv = "^1.0.1"
langchain = "^0.3.3"
langchain-openai = "^0.2.2"

[tool.taskipy.tasks]
format = "ruff check . --fix && black ."

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
